Create README.md
I'm a sophomore student in public university which is Ferdowsi university of Mashhad that is one of the best universities in Iran. Individually I'm a junior researcher in my field of study. I will have an intern in one of the private organization that is related to my interests and my major . After, two years of being student I can claim that I'm an active and eager student for learning HRM and EHRM. Pleas contact me via my Email.